S21/Python_забег/day07/requirements.txt
pipreqs
isort
pylint
autopep8
pytest-cov
sphinx
sphinx-rtd-theme
sphinx-autodoc-typehints